DELIA-40553: Duplicate timedMetadata reported and wrong time value
Reason for change: During linear playlist refresh , already reported Metadata
was getting reported .Also intermittenly time of Metadata was showing outside
culling window. Fixed the duplicate handling across playlist refresh.On underflow
retune, culled value was resetted to 0 causing timed Metadata time to reset.
Test Procedure: Tested with Peacock linear content and checked Metadata
Risks: Low

+++ jsbindings/jsutils.cpp
@@ -388,7 +388,7 @@ const char* aampPlayer_getNameFromEventType(AAMPEventType type)
  * @param[in] durationMS duration of ad break if its a adbreak metadata
  * @retval JSObject of TimedMetadata generated
  */
-JSObjectRef aamp_CreateTimedMetadataJSObject(JSContextRef context, double timeMS, const char* szName, const char* szContent, const char* id, double durationMS)
+JSObjectRef aamp_CreateTimedMetadataJSObject(JSContextRef context, long long timeMS, const char* szName, const char* szContent, const char* id, double durationMS)
 {
 	JSStringRef name;
